updated average entry calculator js and position size js for logic update

+++ static/js/average_entry.js
@@ -184,7 +184,7 @@ document.addEventListener("DOMContentLoaded", function () {
 
 
             const enteredQuantity =
-                parseFloat(
+                parseInt(
                     row.querySelector(
                         ".avg-quantity"
                     ).value
